ss: replace granularity with checkpoint attribute and add synchronization
This patch creates an attribute for the pss to keep track of the
checkpoints. This process was initially put forth by Matthew Khouzam,
Fabien Reumont-Locke improved it.

Previously, to avoid issues with multithread access, a latch was used for
the checkpoints attribute with the waitForCheckpoints method. To replace
this mechanism, synchronization was added on the methods that modify the
transient state.

Change-Id: I9af752ae154fceb1c31ed57b64f0c17650c55703
Signed-off-by: Abdellah Rahmani <rahmaniabdellah1994@gmail.com>
Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>