tmf.core: Propagate the state provider getter
The partial state system needs to access the same state
provider as the analysis. The state provider is not
guaranteed to be stateless. In fact, it is often stateful.
This patch makes it clear that the same state provider is
provided by the analysis if getOrCreate is called subsequent
times

[Added] Propagate the state provider getter to the interface

Change-Id: Ib38fdddaca647ba173dee4c788c5f7edc0ec4153
Signed-off-by: Abdellah Rahmani <rahmaniabdellah1994@gmail.com>
Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>