profiling: fix flamegraph clear bug
When using the selection flamegraph view, the data provider used is the
same for both views (std and selection) and the view elements would get
cleared. This reset of tree elements would effectively disable the
flamegraph view. This patch adds an additional string to the secondaryId
that is used to fetch the data provider, so that two separate data
providers are used.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>